refactor: cambiar Vehicle.currentMileage de Int a Long
- Vehicle.currentMileage, VehicleEntity.currentMileage y updateCurrentMileage en DAO/repositorios ahora son Long.
- VehicleRules.isMileageValid y SaveVehicleUseCase usan Long para validar y guardar kilometraje inicial.
- MileageUseCases compara lecturas Int contra Vehicle.currentMileage Long y actualiza el vehículo con Long.
- MaintenanceViewModel y NavGraph convierten Vehicle.currentMileage a Int solo para los módulos de mantenimiento existentes.
- MaintenanceWorker calcula con Long para evitar desbordamientos.
- Tests y MockDataFactory actualizados con literales L.
- Verificación: assembleDebug + testDebugUnitTest + lintDebug exitosos.

+++ app/src/main/java/com/micarro/core/navigation/NavGraph.kt
@@ -110,7 +110,7 @@ fun MiCarroNavGraph(
                             Screen.ServiceForm.createRoute(
                                 vehicleId = vehicle.id.toString(),
                                 planId = planId,
-                                lastMileage = vehicle.currentMileage
+                                lastMileage = vehicle.currentMileage.toInt()
                             )
                         )
                     }
